requirements.txt
asgiref==3.5.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.4
django-crispy-forms==1.14.0
django-grappelli==3.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
idna==3.3
openpyxl==3.0.9
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
